refactor: make subdataset treatment standalone
This removes establishing super-subdataset relation from load_tabby,
and moves some shared code to utils file.

This brings additional step to workflow, but makes it easier to
combine tabby and non-tabby updates to project datasets.
